+++ lib/src/widgets/empty.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/widgets.dart';
+
+class Empty extends StatelessWidget {
+  const Empty();
+
+  @override
+  Widget build(BuildContext context) => const SizedBox.shrink();
+}